requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
cryptography>=42.0.0
sqlalchemy-utils>=0.41.1
flask_login
